+++ CHANGES.md
@@ -9,6 +9,7 @@
 * Renamed ImageIoProcessor to Java2dProcessor.
 * Java2dProcessor supports pyramidal TIFF.
 * Improved the thread-safety of FilesystemCache.
+* Improved the security of FilesystemResolver.
 * Source images are streamed directly rather than being re-encoded when the
   IIIF parameters request a full-size, unmodified image.
 * Cantaloupe is now distributed as a .zip archive.